+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- markdownlint-disable -->
 ## [Unreleased]
 
+### Added
+- `<slot name="toolbar">` to `<leaflet-map>`
+
 ## [v1.0.7] - 2020-07-20
 
 ### Fixed